feat: 제휴 장소 단건 조회(지소ì도에서 핀 선택 시) 기능 구현
